[mod]: divided cre exection to separate class
