+++ src/relayer/services/index.ts
@@ -1,3 +1,4 @@
 export * from './manager.provider';
 export * from './context.provider';
 export * from './job-queue.service';
+export * from './cre-executor.service';
