+++ src/types/cre.ts
@@ -1,6 +1,17 @@
 import { Hex } from 'viem';
 
 export namespace CRE {
+    // @todo: decouple native CRE types & input for different methods
+    export type Body<T> = {
+        id: string;
+        jsonrpc: '2.0';
+        method: 'workflows.execute';
+        params: {
+            input: T;
+            workflow: { workflowID: string };
+        };
+    };
+
     export type MessageId = Hex;
     export type Proofs = Hex[]; // bytes32[]
