+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './AppError';
-export * from './createCREJwt';
 export * from './configureDotEnv';
 export * from './object';
 export * from './array';
